Missed some small changes
